Factor out `ShaderPass` from `OpenGlRenderer`
